moved components to post/comment
